# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Server Levels+ Dashboard API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export interface AuthUser {
  id: string;
  username: string;
  discriminator: string;
  /** @nullable */
  avatar?: string | null;
  avatarUrl: string;
}

export interface Guild {
  id: string;
  name: string;
  /** @nullable */
  icon?: string | null;
  /** @nullable */
  iconUrl?: string | null;
  memberCount: number;
  botPresent: boolean;
  isAdmin: boolean;
}

export interface GuildStats {
  totalMembers: number;
  activeToday: number;
  messagesThisWeek: number;
  voiceMinutesThisWeek: number;
  totalXpAwarded: number;
  levelUpsThisWeek: number;
  topLevel: number;
  avgLevel: number;
}

export interface GuildSettings {
  guildId: string;
  xpPerMessage: number;
  xpCooldown: number;
  xpMultiplier: number;
  maxLevel: number;
  dailyXpCap: number;
  /** @nullable */
  announceChannel?: string | null;
  announceMessage: string;
  stackRoles: boolean;
  ignoredChannels: string[];
  ignoredRoles: string[];
  voiceXpEnabled: boolean;
  coinsEnabled: boolean;
  prestigeEnabled: boolean;
}

export interface GuildSettingsInput {
  xpPerMessage?: number;
  xpCooldown?: number;
  xpMultiplier?: number;
  maxLevel?: number;
  dailyXpCap?: number;
  /** @nullable */
  announceChannel?: string | null;
  announceMessage?: string;
  stackRoles?: boolean;
  ignoredChannels?: string[];
  ignoredRoles?: string[];
  voiceXpEnabled?: boolean;
  coinsEnabled?: boolean;
  prestigeEnabled?: boolean;
}

export interface LeaderboardEntry {
  rank: number;
  userId: string;
  username: string;
  discriminator: string;
  /** @nullable */
  avatarUrl?: string | null;
  level: number;
  xp: number;
  totalXp: number;
  messages: number;
  voiceMinutes: number;
  coins: number;
  prestige: number;
}

export interface LeaderboardPage {
  entries: LeaderboardEntry[];
  total: number;
  page: number;
  totalPages: number;
}

export interface Achievement {
  id: string;
  name: string;
  description: string;
  icon: string;
  /** @nullable */
  unlockedAt?: string | null;
}

export interface MemberProfile {
  userId: string;
  guildId: string;
  username: string;
  discriminator: string;
  /** @nullable */
  avatarUrl?: string | null;
  level: number;
  xp: number;
  totalXp: number;
  xpNeeded: number;
  rank: number;
  messages: number;
  voiceMinutes: number;
  coins: number;
  prestige: number;
  joinedAt?: string;
  /** @nullable */
  bio?: string | null;
  achievements: Achievement[];
  badges: string[];
}

export interface MemberPage {
  members: MemberProfile[];
  total: number;
  page: number;
  totalPages: number;
}

export interface LevelRole {
  guildId: string;
  level: number;
  roleId: string;
  roleName: string;
}

export interface LevelRoleInput {
  level: number;
  roleId: string;
}

export type XpModificationAction = typeof XpModificationAction[keyof typeof XpModificationAction];


export const XpModificationAction = {
  add: 'add',
  remove: 'remove',
  set: 'set',
} as const;

export interface XpModification {
  amount: number;
  action: XpModificationAction;
}

export type ActivityEventType = typeof ActivityEventType[keyof typeof ActivityEventType];


export const ActivityEventType = {
  level_up: 'level_up',
  achievement: 'achievement',
  prestige: 'prestige',
  daily_reward: 'daily_reward',
  weekly_reward: 'weekly_reward',
} as const;

export interface ActivityEvent {
  id: string;
  type: ActivityEventType;
  userId: string;
  username: string;
  /** @nullable */
  avatarUrl?: string | null;
  /** @nullable */
  level?: number | null;
  /** @nullable */
  achievement?: string | null;
  timestamp: string;
}

export type ShopItemCategory = typeof ShopItemCategory[keyof typeof ShopItemCategory];


export const ShopItemCategory = {
  background: 'background',
  frame: 'frame',
  color: 'color',
  boost: 'boost',
  decoration: 'decoration',
} as const;

export interface ShopItem {
  id: string;
  name: string;
  description: string;
  category: ShopItemCategory;
  price: number;
  icon: string;
}

export type GetLeaderboardParams = {
type?: GetLeaderboardType;
page?: number;
limit?: number;
};

export type GetLeaderboardType = typeof GetLeaderboardType[keyof typeof GetLeaderboardType];


export const GetLeaderboardType = {
  xp: 'xp',
  messages: 'messages',
  voice: 'voice',
  coins: 'coins',
  weekly: 'weekly',
  monthly: 'monthly',
} as const;

export type GetGuildMembersParams = {
search?: string;
page?: number;
};

export type GetGuildActivityParams = {
limit?: number;
};