# lib/api-zod/src/generated/types/achievement.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Server Levels+ Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface Achievement {
  id: string;
  name: string;
  description: string;
  icon: string;
  /** @nullable */
  unlockedAt?: string | null;
}