# lib/api-zod/src/generated/types/activityEvent.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Server Levels+ Dashboard API
 * OpenAPI spec version: 0.1.0
 */
import type { ActivityEventType } from './activityEventType';

export interface ActivityEvent {
  id: string;
  type: ActivityEventType;
  userId: string;
  username: string;
  /** @nullable */
  avatarUrl?: string | null;
  /** @nullable */
  level?: number | null;
  /** @nullable */
  achievement?: string | null;
  timestamp: string;
}