# lib/api-zod/src/generated/types/guildSettingsInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Server Levels+ Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface GuildSettingsInput {
  xpPerMessage?: number;
  xpCooldown?: number;
  xpMultiplier?: number;
  maxLevel?: number;
  dailyXpCap?: number;
  /** @nullable */
  announceChannel?: string | null;
  announceMessage?: string;
  stackRoles?: boolean;
  ignoredChannels?: string[];
  ignoredRoles?: string[];
  voiceXpEnabled?: boolean;
  coinsEnabled?: boolean;
  prestigeEnabled?: boolean;
}