# lib/api-zod/src/generated/types/guildStats.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Server Levels+ Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface GuildStats {
  totalMembers: number;
  activeToday: number;
  messagesThisWeek: number;
  voiceMinutesThisWeek: number;
  totalXpAwarded: number;
  levelUpsThisWeek: number;
  topLevel: number;
  avgLevel: number;
}