# lib/api-zod/src/generated/types/leaderboardPage.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Server Levels+ Dashboard API
 * OpenAPI spec version: 0.1.0
 */
import type { LeaderboardEntry } from './leaderboardEntry';

export interface LeaderboardPage {
  entries: LeaderboardEntry[];
  total: number;
  page: number;
  totalPages: number;
}