# lib/api-zod/src/generated/types/levelRole.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Server Levels+ Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface LevelRole {
  guildId: string;
  level: number;
  roleId: string;
  roleName: string;
}