# lib/api-zod/src/generated/types/xpModification.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Server Levels+ Dashboard API
 * OpenAPI spec version: 0.1.0
 */
import type { XpModificationAction } from './xpModificationAction';

export interface XpModification {
  amount: number;
  action: XpModificationAction;
}